use size_t or most metrics

+++ bench/bench-lcs.cpp
@@ -120,7 +120,7 @@ static void BM_LCS_SIMD(benchmark::State& state)
 {
     std::vector<std::string> seq1;
     std::vector<std::string> seq2;
-    std::vector<int64_t> results(32 * 3 * 4);
+    std::vector<size_t> results(32 * 3 * 4);
     for (int i = 0; i < 32 * 3 * 4; i++)
         seq1.push_back(generate(MaxLen));
     for (int i = 0; i < 10000; i++)
